more matrix multiplication cleanup and improvements
